Add some documentation

+++ src/main/scala/analysis/Interpreter.scala
@@ -2,10 +2,15 @@ package analysis
 
 import cfg._
 
-class Interpreter(fun: Function) extends ValueAnalysis[Integer](fun) {
+/**
+ * An Interpreter for the SSA-CFG format that implicitly defines its semantics.
+ */
+class Interpreter(fun: Function) extends ValueAnalysis[BigInt](fun) {
   override def run(): Unit = {
     populateSymbolTable()
     var currBB = fun.First
+
+
 
   }
